selenium-python/selenium-python-main/requirements.txt
selenium===4.19.0
pytest===8.1.1
allure-pytest===2.13.5
ddt~=1.7.2